requirements.txt
SpeechRecognition==3.10.1
google-generativeai==0.3.2
pyttsx3==2.90
python-dotenv==1.0.0
opencv-python==4.8.1.78
pillow==10.1.0
tkinter==0.1.0